Time: 20 ms (44.41%), Space: 46.1 MB (8.94%) - LeetHub

+++ 0907-koko-eating-bananas/0907-koko-eating-bananas.java
@@ -19,7 +19,7 @@ class Solution {
     public static boolean canEat(int[] piles ,int h,int k){
         int hours_needed=0;
         for(int pile:piles){
-            hours_needed+=Math.ceil(pile/(double)k);
+            hours_needed+=(int)Math.ceil(pile/(double)k);
         }
         return hours_needed<=h;
     }
